Moved formatting logic to Printer class
